+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 - Calendar: Fixed: ErrorException in /usr/share/groupoffice/go/base/util/icalendar/RRuleIterator.php at line 98: Illegal offset type
 - Timeregistration2: fix check on overwriting travel distance for existing time entry
 - Core: Updated German translation
+- Tasks: Clean up task lists when project is being deleted
 
 19-01-2023 6.6.146
 - Email: Fixed vcard import from email